requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==2.10
packet-python==1.43.1
pycparser==2.20
pycrypto==2.6.1
pyghmi==1.5.22
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
